examples/StrawberryExample/pyproject.toml
[tool.poetry]
name = "strawberrytest"
version = "1.0.0"
description = ""
authors = ["Hans Erik Heggem <hanserik@neate.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<=3.11"
fastapi = "^0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.131.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
